Change data types to match DB

+++ DTO/RunStatsDto.cs
@@ -2,10 +2,10 @@
 {
     public class RunStatsDto
     {
-        public required int Type { get; set; }      // 0 = Map, 1 = Bonus, 2 = Stage
+        public required short Type { get; set; }      // 0 = Map, 1 = Bonus, 2 = Stage
         public required int RunTime { get; set; }
-        public required int Stage { get; set; }     // Signifies Bonus number when Type == 1
-        public required int Style { get; set; }
+        public required short Stage { get; set; }     // Signifies Bonus number when Type == 1
+        public required short Style { get; set; }
         public string? Name { get; set; }
         public required float StartVelX { get; set; }
         public required float StartVelY { get; set; }
